asio: Don't add lkl file descriptors to the reactor upon opening
This prevented a socket to be closed and another socket to be opened,
because the file descriptor was still the same but it was not deregistered
properly. Also, waiting on these FDs does not make sense and can be
dangerous, as they start with 0, which would break when detaching from the terminal.